Organize storage of components
Organize components into appropriate folders for better readability and future support
