+++ components/formCards/FormCardsSkeleton.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import { Skeleton } from './ui/skeleton'
+import { Skeleton } from '../ui/skeleton'
 
 function FormCardsSkeleton() {
     return (
