+++ components/navbar/ThemeSwitcher.tsx
@@ -2,7 +2,7 @@
 
 import { useTheme } from 'next-themes';
 import React, { useEffect, useState } from 'react';
-import { Tabs, TabsList, TabsTrigger } from './ui/tabs';
+import { Tabs, TabsList, TabsTrigger } from '../ui/tabs';
 import { DesktopIcon, MoonIcon, SunIcon } from '@radix-ui/react-icons';
 
 function ThemeSwitcher() {
